+++ src/pytest_rerunclassfailures/pytest_rerunclassfailures.py
@@ -7,12 +7,14 @@ from typing import Tuple
 
 import pytest
 import _pytest.nodes
+from _pytest.terminal import TerminalReporter
+from _pytest.config import Config
 from _pytest.reports import TestReport
 from _pytest.runner import runtestprotocol, pytest_runtest_protocol
 from _pytest.config.argparsing import Parser
 
 
-def pytest_addoption(parser: Parser):
+def pytest_addoption(parser: Parser) -> None:
     """
     Add options to the parser.
 
@@ -94,7 +96,9 @@ class RerunClassPlugin:  # pylint: disable=too-few-public-methods
             item.ihook.pytest_runtest_logfinish(nodeid=item.nodeid, location=item.location)
 
     @pytest.hookimpl(tryfirst=True)
-    def pytest_runtest_protocol(self, item: _pytest.nodes.Item, nextitem: _pytest.nodes.Item):  # pylint: disable=W0613
+    def pytest_runtest_protocol(
+        self, item: _pytest.nodes.Item, nextitem: _pytest.nodes.Item  # pylint: disable=W0613
+    ) -> bool:
         """
         Run the test protocol.
 
@@ -102,6 +106,8 @@ class RerunClassPlugin:  # pylint: disable=too-few-public-methods
         :type item: _pytest.nodes.Item
         :param nextitem: next pytest item
         :type nextitem: _pytest.nodes.Item
+        :return: True if any actions of plugin performed, False otherwise
+        :rtype: bool
         """
         parent_class = item.getparent(pytest.Class)
         module = item.nodeid.split("::")[0]
@@ -203,12 +209,12 @@ class RerunClassPlugin:  # pylint: disable=too-few-public-methods
 
         return siblings
 
-    def _save_parent_initial_state(self, parent):
+    def _save_parent_initial_state(self, parent: pytest.Class) -> dict:
         """
         Save the parent initial state.
 
         :param parent: pytest item
-        :type parent: _pytest.
+        :type parent: _pytest.Item
         :return: parent initial state
         :rtype: dict
         """
@@ -331,8 +337,41 @@ class RerunClassPlugin:  # pylint: disable=too-few-public-methods
                             self.logger.debug("Removing cached result for %s", fixture_def)
                             setattr(fixture_def, cached_result, None)
 
+    def pytest_terminal_summary(
+        self, terminalreporter: TerminalReporter, exitstatus: int, config: Config  # pylint: disable=unused-argument
+    ) -> None:
+        """
+        Reports reruns section to terminal.
 
-def pytest_configure(config: pytest.Config):
+        :param terminalreporter: pytest terminal reporter
+        :type terminalreporter: _pytest.terminal.TerminalReporter
+        :param exitstatus: exit status
+        :type exitstatus: int
+        :param config: pytest config
+        :type config: _pytest.config.Config
+        :return: None
+        :rtype: None
+        """
+        if "rerun" not in terminalreporter.stats:
+            return
+
+        terminalreporter._tw.sep("=", "RERUNS")  # pylint: disable=W0212
+
+        for rerun_test in terminalreporter.stats["rerun"]:
+            pos = rerun_test.nodeid
+            terminalreporter._tw.line(f"RERUN {pos}")  # pylint: disable=W0212
+
+            if hasattr(rerun_test, "longrepr"):
+                if isinstance(rerun_test.longrepr, tuple):
+                    for line in rerun_test.longrepr:
+                        if line:
+                            terminalreporter._tw.line(str(line))  # pylint: disable=W0212
+                else:
+                    if rerun_test.longrepr:
+                        terminalreporter._tw.line(str(rerun_test.longrepr))  # pylint: disable=W0212
+
+
+def pytest_configure(config: Config) -> None:
     """
     Configure the plugin.